moving true sim odom to seperate class
