+++ simgui-window.json
@@ -10,8 +10,8 @@
       "style": "0",
       "userScale": "2",
       "width": "1440",
-      "xpos": "80",
-      "ypos": "260"
+      "xpos": "362",
+      "ypos": "271"
     }
   },
   "Table": {
